dssdsaddas

+++ 42_cursus/Libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: olcherno <olcherno@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/01 14:37:56 by olcherno          #+#    #+#             */
-/*   Updated: 2024/10/01 14:56:41 by olcherno         ###   ########.fr       */
+/*   Updated: 2024/10/03 17:44:20 by olcherno         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
